Add test case for edge case
The edge case is a bit silly, but supporting it stops us from choking on
any kind of possible AST node arrangement there, so in a sense, it is
cleaner.

We test the behavior of there being two declarators, one of which is a
component and one that is not and the implication it has on where
existing comments have to move to.